memwebapp/backend/pyproject.toml
[project]
name = "memoapp-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==23.2.1",
    "annotated-types==0.7.0",
    "anyio==3.7.1",
    "bcrypt==4.3.0",
    "boto3>=1.40.31",
    "certifi==2025.7.14",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.1.8",
    "cryptography>=43.0.0,<44.0.0",
    "distro==1.9.0",
    "ecdsa==0.19.1",
    "fastapi==0.104.1",
    "filelock==3.18.0",
    "fsspec==2025.7.0",
    "google-auth==2.23.4",
    "google-auth-httplib2==0.1.1",
    "google-auth-oauthlib==1.1.0",
    "google-generativeai>=0.8.5",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "markupsafe==3.0.2",
    "mpmath==1.3.0",
    "msal==1.24.1",
    "networkx==3.2.1",
    "numpy>=1.24.0",
    "openai==1.3.7",
    "pandas>=2.1.0",
    "passlib==1.7.4",
    "psycopg2-binary>=2.9.0",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pydantic>=2.8.0",
    "pydub==0.25.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "pytz==2025.2",
    "requests==2.32.4",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "speechrecognition==3.10.0",
    "sqlalchemy>=2.0.25",
    "starlette==0.27.0",
    "sympy==1.14.0",
    "tqdm==4.67.1",
    "typing-extensions==4.14.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uv==0.8.4",
    "uvicorn==0.24.0",
    "whisper==1.1.10",
]
